Add a class named UpdateMethodCacher is used for updating or cleaning cache
