+++ src/main/java/com/jcabi/aspects/aj/MethodCacher.java
@@ -32,17 +32,13 @@ package com.jcabi.aspects.aj;
 import com.jcabi.aspects.Cacheable;
 import com.jcabi.aspects.Loggable;
 import com.jcabi.log.Logger;
-import com.jcabi.log.VerboseRunnable;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.util.Arrays;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.Executors;
 import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.ProceedingJoinPoint;
@@ -66,9 +62,7 @@ import org.aspectj.lang.reflect.MethodSignature;
  * @since 0.8
  */
 @Aspect
-@SuppressWarnings(
-    { "PMD.DoNotUseThreads", "PMD.TooManyMethods", "PMD.GodClass" }
-)
+@SuppressWarnings("PMD.TooManyMethods")
 public final class MethodCacher {
 
     /**
@@ -76,62 +70,19 @@ public final class MethodCacher {
      * @checkstyle LineLength (2 lines)
      */
     private final transient ConcurrentMap<MethodCacher.Key, MethodCacher.Tunnel> tunnels =
-        new ConcurrentHashMap<MethodCacher.Key, MethodCacher.Tunnel>(0);
+        new ConcurrentHashMap<Key, Tunnel>(0);
 
     /**
      * Save the keys which need update.
      */
-    private final transient BlockingQueue<MethodCacher.Key> updatekeys =
+    private final transient BlockingQueue<Key> updatekeys =
         new LinkedBlockingQueue<MethodCacher.Key>();
-
-    /**
-     * Service that cleans cache.
-     */
-    private final transient ScheduledExecutorService cleaner =
-        Executors.newSingleThreadScheduledExecutor(
-            new NamedThreads(
-                "cacheable-clean",
-                "automated cleaning of expired @Cacheable values"
-            )
-        );
-
-    /**
-     * Service that update cache.
-     */
-    private final transient ScheduledExecutorService updater =
-        Executors.newSingleThreadScheduledExecutor(
-            new NamedThreads(
-                "cacheable-update",
-                "async update of expired @Cacheable values"
-            )
-        );
 
     /**
      * Public ctor.
      */
     public MethodCacher() {
-        this.cleaner.scheduleWithFixedDelay(
-            new VerboseRunnable(
-                new Runnable() {
-                    @Override
-                    public void run() {
-                        MethodCacher.this.clean();
-                    }
-                }
-            ),
-            1L, 1L, TimeUnit.SECONDS
-        );
-        this.updater.schedule(
-            new VerboseRunnable(
-                new Runnable() {
-                    @Override
-                    public void run() {
-                        MethodCacher.this.update();
-                    }
-                }
-            ),
-            0L, TimeUnit.SECONDS
-        );
+        new UpdateMethodCacher(this.tunnels, this.updatekeys).start();
     }
 
     /**
@@ -282,65 +233,9 @@ public final class MethodCacher {
     }
 
     /**
-     * Clean cache.
-     */
-    private void clean() {
-        synchronized (this.tunnels) {
-            for (final MethodCacher.Key key : this.tunnels.keySet()) {
-                if (this.tunnels.get(key).expired()
-                    && !this.tunnels.get(key).asyncUpdate()) {
-                    final MethodCacher.Tunnel tunnel = this.tunnels.remove(key);
-                    LogHelper.log(
-                        key.getLevel(),
-                        this,
-                        "%s:%s expired in cache",
-                        key,
-                        tunnel
-                    );
-                }
-            }
-        }
-    }
-
-    /**
-     * Update cache.
-     */
-    @SuppressWarnings("PMD.AvoidCatchingThrowable")
-    private void update() {
-        while (true) {
-            try {
-                final MethodCacher.Key key = this.updatekeys.take();
-                final MethodCacher.Tunnel tunnel = this.tunnels.get(key);
-                if (tunnel != null && tunnel.expired()) {
-                    final MethodCacher.Tunnel newTunnel = tunnel.copy();
-                    newTunnel.through();
-                    this.tunnels.put(key, newTunnel);
-                }
-            } catch (final InterruptedException ex) {
-                LogHelper.log(
-                    Loggable.ERROR,
-                    this,
-                    "%s:%s",
-                    ex.getMessage(),
-                    ex
-                );
-            // @checkstyle IllegalCatch (1 line)
-            } catch (final Throwable ex) {
-                LogHelper.log(
-                    Loggable.ERROR,
-                    this,
-                    "Exception message is %s, Exception is %s",
-                    ex.getMessage(),
-                    ex
-                );
-            }
-        }
-    }
-
-    /**
      * Mutable caching/calling tunnel, it is thread-safe.
      */
-    private static final class Tunnel {
+    protected static final class Tunnel {
         /**
          * Proceeding join point.
          */
@@ -460,7 +355,7 @@ public final class MethodCacher {
     /**
      * Key of a callable target.
      */
-    private static final class Key {
+    protected static final class Key {
         /**
          * When instantiated.
          */
@@ -502,6 +397,15 @@ public final class MethodCacher {
                 this.level = Loggable.DEBUG;
             }
         }
+
+        /**
+         * Get log level.
+         * @return Log level of current method.
+         */
+        public int getLevel() {
+            return this.level;
+        }
+
         @Override
         public String toString() {
             return Mnemos.toText(this.method, this.arguments, true, false);
@@ -571,13 +475,6 @@ public final class MethodCacher {
             return tgt;
         }
 
-        /**
-         * Get log level.
-         * @return Log level of current method.
-         */
-        private int getLevel() {
-            return this.level;
-        }
     }
 
 }
